requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
asgiref==3.3.1
    # via django
attrs==20.3.0
    # via jsonschema
certifi==2020.12.5
    # via requests
chardet==4.0.0
    # via requests
django-cors-headers==3.6.0
    # via -r requirements.in
django-filter==2.4.0
    # via -r requirements.in
django-otp==1.0.2
    # via -r requirements.in
django==3.1.5
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-filter
    #   django-otp
    #   djangorestframework
    #   drf-spectacular
djangorestframework==3.12.2
    # via
    #   -r requirements.in
    #   drf-spectacular
drf-spectacular==0.13.0
    # via -r requirements.in
idna==2.10
    # via requests
inflection==0.5.1
    # via drf-spectacular
jsonschema==3.2.0
    # via drf-spectacular
pillow==8.1.0
    # via -r requirements.in
psycopg2-binary==2.8.6
    # via -r requirements.in
pyrsistent==0.17.3
    # via jsonschema
pytz==2020.5
    # via django
pyyaml==5.3.1
    # via drf-spectacular
requests==2.25.1
    # via -r requirements.in
six==1.15.0
    # via jsonschema
sqlparse==0.4.1
    # via django
uritemplate==3.0.1
    # via drf-spectacular
urllib3==1.26.2
    # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
